refactoring props and types
